Move DeleteBlock to components/common
