--- airbyte-webapp/src/components/DeleteBlock/index.tsx
@@ -1,3 +1,0 @@
-import DeleteBlock from "./DeleteBlock";
-
-export default DeleteBlock;
