+++ airbyte-webapp/src/components/common/DeleteBlock/index.ts
@@ -1,0 +1,1 @@
+export * from "./DeleteBlock";
